+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KAGFMWXRPNFS9M3111A1DQ8R"
+        "01KAGR8S1R1FCTPDHS9Z665RTK"
     )
     r"""The unique identifier of the data source"""