+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -63,5 +63,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-11-20T11:15:48.887Z")
+    created: Optional[datetime] = parse_datetime("2025-11-20T13:46:31.464Z")
     r"""The date and time the resource was created"""