+++ packages/orq-rc/src/orq_ai_sdk/models/listbudgetsop.py
@@ -250,7 +250,7 @@ class ListBudgetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-20T11:15:46.091Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-20T13:46:28.834Z")
     r"""The date and time the resource was last updated"""