+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -130,7 +130,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KAGFMWXMR2Z0GQATTMKNHY4F"
+        "01KAGR8S1KC8YWJAEW0SMPVKAW"
     )
     r"""The unique identifier of the data source"""