+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -80,5 +80,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-20T11:15:45.627Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-20T13:46:28.497Z")
     r"""The date and time the resource was last updated"""