+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KAGFMWXQX7CW0HSTGFDNMJKS"
+        "01KAGR8S1PZ91WS79RDHMAJDN7"
     )
     r"""The unique identifier of the data source"""