+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KAGFMWXT1A2RSD5D8QRA8R7K"
+        "01KAGR8S1V86DJ6YK8850RR3VD"
     )
     r"""The unique identifier of the data source"""