+++ packages/orq-rc/RELEASES.md
@@ -4129,3 +4129,13 @@ Based on:
 - [python v4.0.0-rc.57] packages/orq-rc
 ### Releases
 - [PyPI v4.0.0-rc.57] https://pypi.org/project/orq-ai-sdk/4.0.0-rc.57 - packages/orq-rc
+
+## 2025-11-20 13:48:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.656.1 (2.753.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.0.0-rc.59] packages/orq-rc
+### Releases
+- [PyPI v4.0.0-rc.59] https://pypi.org/project/orq-ai-sdk/4.0.0-rc.59 - packages/orq-rc